refactor: change folder structure for overlay and portal
since they dont properly fit in any folder
