+++ packages/ui-angular/src/index.ts
@@ -1,4 +1,4 @@
 export * from './lib/components';
-export * from './lib/directives';
-export * from './lib/services';
+export * from './lib/overlay';
+export * from './lib/portal';
 export * from './lib/utils';
