--- packages/ui-angular/src/lib/directives/index.ts
@@ -1,3 +1,0 @@
-export * from './portal/portal.component';
-export * from './portal/portal.directive';
-export * from './portal/portal.service';
