+++ packages/ui-angular/src/lib/overlay/floating-adapter.ts
