+++ packages/ui-angular/src/lib/overlay/index.ts
@@ -1,0 +1,1 @@
+export * from './overlay.service';
