+++ packages/ui-angular/src/lib/portal/index.ts
@@ -1,0 +1,3 @@
+export * from './portal.component';
+export * from './portal.directive';
+export * from './portal.service';
