--- packages/ui-angular/src/lib/services/index.ts
@@ -1,1 +1,0 @@
-export * from './overlay/overlay.service';
